+++ src/ncc/ThirdParty/nikic/PhpParser/Builder.php
@@ -2,12 +2,11 @@
 
 namespace ncc\ThirdParty\nikic\PhpParser;
 
-interface Builder
-{
+interface Builder {
     /**
      * Returns the built node.
      *
      * @return Node The built node
      */
-    public function getNode() : Node;
+    public function getNode(): Node;
 }