+++ src/ncc/ThirdParty/nikic/PhpParser/Builder/Declaration.php
@@ -5,16 +5,23 @@ namespace ncc\ThirdParty\nikic\PhpParser\Builder;
 use ncc\ThirdParty\nikic\PhpParser;
 use PhpParser\BuilderHelpers;
 
-abstract class Declaration implements PhpParser\Builder
-{
-    protected $attributes = [];
+abstract class Declaration implements PhpParser\Builder {
+    /** @var array<string, mixed> */
+    protected array $attributes = [];
 
+    /**
+     * Adds a statement.
+     *
+     * @param \ncc\ThirdParty\nikic\PhpParser\Node\Stmt|PhpParser\Builder $stmt The statement to add
+     *
+     * @return $this The builder instance (for fluid interface)
+     */
     abstract public function addStmt($stmt);
 
     /**
      * Adds multiple statements.
      *
-     * @param array $stmts The statements to add
+     * @param (\ncc\ThirdParty\nikic\PhpParser\Node\Stmt|PhpParser\Builder)[] $stmts The statements to add
      *
      * @return $this The builder instance (for fluid interface)
      */