+++ src/ncc/ThirdParty/nikic/PhpParser/Builder/Function_.php
@@ -4,16 +4,16 @@ namespace ncc\ThirdParty\nikic\PhpParser\Builder;
 
 use ncc\ThirdParty\nikic\PhpParser;
 use PhpParser\BuilderHelpers;
-use PhpParser\Node;
-use PhpParser\Node\Stmt;
+use ncc\ThirdParty\nikic\PhpParser\Node;
+use ncc\ThirdParty\nikic\PhpParser\Node\Stmt;
 
-class Function_ extends FunctionLike
-{
-    protected $name;
-    protected $stmts = [];
+class Function_ extends FunctionLike {
+    protected string $name;
+    /** @var list<Stmt> */
+    protected array $stmts = [];
 
-    /** @var Node\AttributeGroup[] */
-    protected $attributeGroups = [];
+    /** @var list<Node\AttributeGroup> */
+    protected array $attributeGroups = [];
 
     /**
      * Creates a function builder.
@@ -55,7 +55,7 @@ class Function_ extends FunctionLike
      *
      * @return Stmt\Function_ The built function node
      */
-    public function getNode() : Node {
+    public function getNode(): Node {
         return new Stmt\Function_($this->name, [
             'byRef'      => $this->returnByRef,
             'params'     => $this->params,