+++ src/ncc/ThirdParty/nikic/PhpParser/Builder/TraitUse.php
@@ -7,10 +7,11 @@ use ncc\ThirdParty\nikic\PhpParser\BuilderHelpers;
 use ncc\ThirdParty\nikic\PhpParser\Node;
 use ncc\ThirdParty\nikic\PhpParser\Node\Stmt;
 
-class TraitUse implements Builder
-{
-    protected $traits = [];
-    protected $adaptations = [];
+class TraitUse implements Builder {
+    /** @var Node\Name[] */
+    protected array $traits = [];
+    /** @var Stmt\TraitUseAdaptation[] */
+    protected array $adaptations = [];
 
     /**
      * Creates a trait use builder.
@@ -58,7 +59,7 @@ class TraitUse implements Builder
      *
      * @return Node The built node
      */
-    public function getNode() : Node {
+    public function getNode(): Node {
         return new Stmt\TraitUse($this->traits, $this->adaptations);
     }
 }