+++ src/ncc/ThirdParty/nikic/PhpParser/BuilderFactory.php
@@ -10,18 +10,15 @@ use ncc\ThirdParty\nikic\PhpParser\Node\Name;
 use ncc\ThirdParty\nikic\PhpParser\Node\Scalar\String_;
 use ncc\ThirdParty\nikic\PhpParser\Node\Stmt\Use_;
 
-class BuilderFactory
-{
+class BuilderFactory {
     /**
      * Creates an attribute node.
      *
      * @param string|Name $name Name of the attribute
-     * @param array       $args Attribute named arguments
-     *
-     * @return Node\Attribute
+     * @param array $args Attribute named arguments
      */
-    public function attribute($name, array $args = []) : Node\Attribute {
-        return new Node\Attribute(
+    public function attribute($name, array $args = []): ncc\ThirdParty\nikic\PhpParser\Node\Attribute {
+        return new ncc\ThirdParty\nikic\PhpParser\Node\Attribute(
             BuilderHelpers::normalizeName($name),
             $this->args($args)
         );
@@ -30,12 +27,12 @@ class BuilderFactory
     /**
      * Creates a namespace builder.
      *
-     * @param null|string|Node\Name $name Name of the namespace
+     * @param null|string|ncc\ThirdParty\nikic\PhpParser\Node\Name $name Name of the namespace
      *
-     * @return Builder\Namespace_ The created namespace builder
+     * @return ncc\ThirdParty\nikic\PhpParser\Builder\Namespace_ The created namespace builder
      */
-    public function namespace($name) : Builder\Namespace_ {
-        return new Builder\Namespace_($name);
+    public function namespace($name): ncc\ThirdParty\nikic\PhpParser\Builder\Namespace_ {
+        return new ncc\ThirdParty\nikic\PhpParser\Builder\Namespace_($name);
     }
 
     /**
@@ -43,10 +40,10 @@ class BuilderFactory
      *
      * @param string $name Name of the class
      *
-     * @return Builder\Class_ The created class builder
+     * @return ncc\ThirdParty\nikic\PhpParser\Builder\Class_ The created class builder
      */
-    public function class(string $name) : Builder\Class_ {
-        return new Builder\Class_($name);
+    public function class(string $name): ncc\ThirdParty\nikic\PhpParser\Builder\Class_ {
+        return new ncc\ThirdParty\nikic\PhpParser\Builder\Class_($name);
     }
 
     /**
@@ -54,10 +51,10 @@ class BuilderFactory
      *
      * @param string $name Name of the interface
      *
-     * @return Builder\Interface_ The created interface builder
+     * @return ncc\ThirdParty\nikic\PhpParser\Builder\Interface_ The created interface builder
      */
-    public function interface(string $name) : Builder\Interface_ {
-        return new Builder\Interface_($name);
+    public function interface(string $name): ncc\ThirdParty\nikic\PhpParser\Builder\Interface_ {
+        return new ncc\ThirdParty\nikic\PhpParser\Builder\Interface_($name);
     }
 
     /**
@@ -65,10 +62,10 @@ class BuilderFactory
      *
      * @param string $name Name of the trait
      *
-     * @return Builder\Trait_ The created trait builder
+     * @return ncc\ThirdParty\nikic\PhpParser\Builder\Trait_ The created trait builder
      */
-    public function trait(string $name) : Builder\Trait_ {
-        return new Builder\Trait_($name);
+    public function trait(string $name): ncc\ThirdParty\nikic\PhpParser\Builder\Trait_ {
+        return new ncc\ThirdParty\nikic\PhpParser\Builder\Trait_($name);
     }
 
     /**
@@ -76,38 +73,38 @@ class BuilderFactory
      *
      * @param string $name Name of the enum
      *
-     * @return Builder\Enum_ The created enum builder
+     * @return ncc\ThirdParty\nikic\PhpParser\Builder\Enum_ The created enum builder
      */
-    public function enum(string $name) : Builder\Enum_ {
-        return new Builder\Enum_($name);
+    public function enum(string $name): ncc\ThirdParty\nikic\PhpParser\Builder\Enum_ {
+        return new ncc\ThirdParty\nikic\PhpParser\Builder\Enum_($name);
     }
 
     /**
      * Creates a trait use builder.
      *
-     * @param Node\Name|string ...$traits Trait names
+     * @param ncc\ThirdParty\nikic\PhpParser\Node\Name|string ...$traits Trait names
      *
-     * @return Builder\TraitUse The create trait use builder
+     * @return ncc\ThirdParty\nikic\PhpParser\Builder\TraitUse The created trait use builder
      */
-    public function useTrait(...$traits) : Builder\TraitUse {
-        return new Builder\TraitUse(...$traits);
+    public function useTrait(...$traits): ncc\ThirdParty\nikic\PhpParser\Builder\TraitUse {
+        return new ncc\ThirdParty\nikic\PhpParser\Builder\TraitUse(...$traits);
     }
 
     /**
      * Creates a trait use adaptation builder.
      *
-     * @param Node\Name|string|null  $trait  Trait name
-     * @param Node\Identifier|string $method Method name
+     * @param ncc\ThirdParty\nikic\PhpParser\Node\Name|string|null $trait Trait name
+     * @param ncc\ThirdParty\nikic\PhpParser\Node\Identifier|string $method Method name
      *
-     * @return Builder\TraitUseAdaptation The create trait use adaptation builder
+     * @return ncc\ThirdParty\nikic\PhpParser\Builder\TraitUseAdaptation The created trait use adaptation builder
      */
-    public function traitUseAdaptation($trait, $method = null) : Builder\TraitUseAdaptation {
+    public function traitUseAdaptation($trait, $method = null): ncc\ThirdParty\nikic\PhpParser\Builder\TraitUseAdaptation {
         if ($method === null) {
             $method = $trait;
             $trait = null;
         }
 
-        return new Builder\TraitUseAdaptation($trait, $method);
+        return new ncc\ThirdParty\nikic\PhpParser\Builder\TraitUseAdaptation($trait, $method);
     }
 
     /**
@@ -115,10 +112,10 @@ class BuilderFactory
      *
      * @param string $name Name of the method
      *
-     * @return Builder\Method The created method builder
+     * @return ncc\ThirdParty\nikic\PhpParser\Builder\Method The created method builder
      */
-    public function method(string $name) : Builder\Method {
-        return new Builder\Method($name);
+    public function method(string $name): ncc\ThirdParty\nikic\PhpParser\Builder\Method {
+        return new ncc\ThirdParty\nikic\PhpParser\Builder\Method($name);
     }
 
     /**
@@ -126,10 +123,10 @@ class BuilderFactory
      *
      * @param string $name Name of the parameter
      *
-     * @return Builder\Param The created parameter builder
+     * @return ncc\ThirdParty\nikic\PhpParser\Builder\Param The created parameter builder
      */
-    public function param(string $name) : Builder\Param {
-        return new Builder\Param($name);
+    public function param(string $name): ncc\ThirdParty\nikic\PhpParser\Builder\Param {
+        return new ncc\ThirdParty\nikic\PhpParser\Builder\Param($name);
     }
 
     /**
@@ -137,10 +134,10 @@ class BuilderFactory
      *
      * @param string $name Name of the property
      *
-     * @return Builder\Property The created property builder
+     * @return ncc\ThirdParty\nikic\PhpParser\Builder\Property The created property builder
      */
-    public function property(string $name) : Builder\Property {
-        return new Builder\Property($name);
+    public function property(string $name): ncc\ThirdParty\nikic\PhpParser\Builder\Property {
+        return new ncc\ThirdParty\nikic\PhpParser\Builder\Property($name);
     }
 
     /**
@@ -148,76 +145,74 @@ class BuilderFactory
      *
      * @param string $name Name of the function
      *
-     * @return Builder\Function_ The created function builder
+     * @return ncc\ThirdParty\nikic\PhpParser\Builder\Function_ The created function builder
      */
-    public function function(string $name) : Builder\Function_ {
-        return new Builder\Function_($name);
+    public function function(string $name): ncc\ThirdParty\nikic\PhpParser\Builder\Function_ {
+        return new ncc\ThirdParty\nikic\PhpParser\Builder\Function_($name);
     }
 
     /**
      * Creates a namespace/class use builder.
      *
-     * @param Node\Name|string $name Name of the entity (namespace or class) to alias
+     * @param ncc\ThirdParty\nikic\PhpParser\Node\Name|string $name Name of the entity (namespace or class) to alias
      *
-     * @return Builder\Use_ The created use builder
+     * @return ncc\ThirdParty\nikic\PhpParser\Builder\Use_ The created use builder
      */
-    public function use($name) : Builder\Use_ {
-        return new Builder\Use_($name, Use_::TYPE_NORMAL);
+    public function use($name): ncc\ThirdParty\nikic\PhpParser\Builder\Use_ {
+        return new ncc\ThirdParty\nikic\PhpParser\Builder\Use_($name, Use_::TYPE_NORMAL);
     }
 
     /**
      * Creates a function use builder.
      *
-     * @param Node\Name|string $name Name of the function to alias
+     * @param ncc\ThirdParty\nikic\PhpParser\Node\Name|string $name Name of the function to alias
      *
-     * @return Builder\Use_ The created use function builder
+     * @return ncc\ThirdParty\nikic\PhpParser\Builder\Use_ The created use function builder
      */
-    public function useFunction($name) : Builder\Use_ {
-        return new Builder\Use_($name, Use_::TYPE_FUNCTION);
+    public function useFunction($name): ncc\ThirdParty\nikic\PhpParser\Builder\Use_ {
+        return new ncc\ThirdParty\nikic\PhpParser\Builder\Use_($name, Use_::TYPE_FUNCTION);
     }
 
     /**
      * Creates a constant use builder.
      *
-     * @param Node\Name|string $name Name of the const to alias
+     * @param ncc\ThirdParty\nikic\PhpParser\Node\Name|string $name Name of the const to alias
      *
-     * @return Builder\Use_ The created use const builder
+     * @return ncc\ThirdParty\nikic\PhpParser\Builder\Use_ The created use const builder
      */
-    public function useConst($name) : Builder\Use_ {
-        return new Builder\Use_($name, Use_::TYPE_CONSTANT);
+    public function useConst($name): ncc\ThirdParty\nikic\PhpParser\Builder\Use_ {
+        return new ncc\ThirdParty\nikic\PhpParser\Builder\Use_($name, Use_::TYPE_CONSTANT);
     }
 
     /**
      * Creates a class constant builder.
      *
-     * @param string|Identifier                          $name  Name
-     * @param Node\Expr|bool|null|int|float|string|array $value Value
+     * @param string|Identifier $name Name
+     * @param ncc\ThirdParty\nikic\PhpParser\Node\Expr|bool|null|int|float|string|array $value Value
      *
-     * @return Builder\ClassConst The created use const builder
+     * @return ncc\ThirdParty\nikic\PhpParser\Builder\ClassConst The created use const builder
      */
-    public function classConst($name, $value) : Builder\ClassConst {
-        return new Builder\ClassConst($name, $value);
+    public function classConst($name, $value): ncc\ThirdParty\nikic\PhpParser\Builder\ClassConst {
+        return new ncc\ThirdParty\nikic\PhpParser\Builder\ClassConst($name, $value);
     }
 
     /**
      * Creates an enum case builder.
      *
-     * @param string|Identifier $name  Name
+     * @param string|Identifier $name Name
      *
-     * @return Builder\EnumCase The created use const builder
+     * @return ncc\ThirdParty\nikic\PhpParser\Builder\EnumCase The created use const builder
      */
-    public function enumCase($name) : Builder\EnumCase {
-        return new Builder\EnumCase($name);
+    public function enumCase($name): ncc\ThirdParty\nikic\PhpParser\Builder\EnumCase {
+        return new ncc\ThirdParty\nikic\PhpParser\Builder\EnumCase($name);
     }
 
     /**
      * Creates node a for a literal value.
      *
-     * @param Expr|bool|null|int|float|string|array $value $value
-     *
-     * @return Expr
+     * @param Expr|bool|null|int|float|string|array|\UnitEnum $value $value
      */
-    public function val($value) : Expr {
+    public function val($value): Expr {
         return BuilderHelpers::normalizeValue($value);
     }
 
@@ -225,10 +220,8 @@ class BuilderFactory
      * Creates variable node.
      *
      * @param string|Expr $name Name
-     *
-     * @return Expr\Variable
      */
-    public function var($name) : Expr\Variable {
+    public function var($name): Expr\Variable {
         if (!\is_string($name) && !$name instanceof Expr) {
             throw new \LogicException('Variable name must be string or Expr');
         }
@@ -243,9 +236,9 @@ class BuilderFactory
      *
      * @param array $args List of arguments to normalize
      *
-     * @return Arg[]
+     * @return list<Arg>
      */
-    public function args(array $args) : array {
+    public function args(array $args): array {
         $normalizedArgs = [];
         foreach ($args as $key => $arg) {
             if (!($arg instanceof Arg)) {
@@ -263,11 +256,9 @@ class BuilderFactory
      * Creates a function call node.
      *
      * @param string|Name|Expr $name Function name
-     * @param array            $args Function arguments
-     *
-     * @return Expr\FuncCall
+     * @param array $args Function arguments
      */
-    public function funcCall($name, array $args = []) : Expr\FuncCall {
+    public function funcCall($name, array $args = []): Expr\FuncCall {
         return new Expr\FuncCall(
             BuilderHelpers::normalizeNameOrExpr($name),
             $this->args($args)
@@ -277,13 +268,11 @@ class BuilderFactory
     /**
      * Creates a method call node.
      *
-     * @param Expr                   $var  Variable the method is called on
+     * @param Expr $var Variable the method is called on
      * @param string|Identifier|Expr $name Method name
-     * @param array                  $args Method arguments
-     *
-     * @return Expr\MethodCall
+     * @param array $args Method arguments
      */
-    public function methodCall(Expr $var, $name, array $args = []) : Expr\MethodCall {
+    public function methodCall(Expr $var, $name, array $args = []): Expr\MethodCall {
         return new Expr\MethodCall(
             $var,
             BuilderHelpers::normalizeIdentifierOrExpr($name),
@@ -294,13 +283,11 @@ class BuilderFactory
     /**
      * Creates a static method call node.
      *
-     * @param string|Name|Expr       $class Class name
-     * @param string|Identifier|Expr $name  Method name
-     * @param array                  $args  Method arguments
-     *
-     * @return Expr\StaticCall
+     * @param string|Name|Expr $class Class name
+     * @param string|Identifier|Expr $name Method name
+     * @param array $args Method arguments
      */
-    public function staticCall($class, $name, array $args = []) : Expr\StaticCall {
+    public function staticCall($class, $name, array $args = []): Expr\StaticCall {
         return new Expr\StaticCall(
             BuilderHelpers::normalizeNameOrExpr($class),
             BuilderHelpers::normalizeIdentifierOrExpr($name),
@@ -312,11 +299,9 @@ class BuilderFactory
      * Creates an object creation node.
      *
      * @param string|Name|Expr $class Class name
-     * @param array            $args  Constructor arguments
-     *
-     * @return Expr\New_
+     * @param array $args Constructor arguments
      */
-    public function new($class, array $args = []) : Expr\New_ {
+    public function new($class, array $args = []): Expr\New_ {
         return new Expr\New_(
             BuilderHelpers::normalizeNameOrExpr($class),
             $this->args($args)
@@ -327,22 +312,18 @@ class BuilderFactory
      * Creates a constant fetch node.
      *
      * @param string|Name $name Constant name
-     *
-     * @return Expr\ConstFetch
      */
-    public function constFetch($name) : Expr\ConstFetch {
+    public function constFetch($name): Expr\ConstFetch {
         return new Expr\ConstFetch(BuilderHelpers::normalizeName($name));
     }
 
     /**
      * Creates a property fetch node.
      *
-     * @param Expr                   $var  Variable holding object
+     * @param Expr $var Variable holding object
      * @param string|Identifier|Expr $name Property name
-     *
-     * @return Expr\PropertyFetch
      */
-    public function propertyFetch(Expr $var, $name) : Expr\PropertyFetch {
+    public function propertyFetch(Expr $var, $name): Expr\PropertyFetch {
         return new Expr\PropertyFetch($var, BuilderHelpers::normalizeIdentifierOrExpr($name));
     }
 
@@ -350,9 +331,7 @@ class BuilderFactory
      * Creates a class constant fetch node.
      *
      * @param string|Name|Expr $class Class name
-     * @param string|Identifier|Expr $name  Constant name
-     *
-     * @return Expr\ClassConstFetch
+     * @param string|Identifier|Expr $name Constant name
      */
     public function classConstFetch($class, $name): Expr\ClassConstFetch {
         return new Expr\ClassConstFetch(
@@ -365,10 +344,8 @@ class BuilderFactory
      * Creates nested Concat nodes from a list of expressions.
      *
      * @param Expr|string ...$exprs Expressions or literal strings
-     *
-     * @return Concat
      */
-    public function concat(...$exprs) : Concat {
+    public function concat(...$exprs): Concat {
         $numExprs = count($exprs);
         if ($numExprs < 2) {
             throw new \LogicException('Expected at least two expressions');
@@ -383,9 +360,8 @@ class BuilderFactory
 
     /**
      * @param string|Expr $expr
-     * @return Expr
      */
-    private function normalizeStringExpr($expr) : Expr {
+    private function normalizeStringExpr($expr): Expr {
         if ($expr instanceof Expr) {
             return $expr;
         }