+++ src/ncc/ThirdParty/nikic/PhpParser/Lexer/TokenEmulator/MatchTokenEmulator.php
@@ -2,22 +2,18 @@
 
 namespace ncc\ThirdParty\nikic\PhpParser\Lexer\TokenEmulator;
 
-use ncc\ThirdParty\nikic\PhpParser\Lexer\Emulative;
+use ncc\ThirdParty\nikic\PhpParser\PhpVersion;
 
-final class MatchTokenEmulator extends KeywordEmulator
-{
-    public function getPhpVersion(): string
-    {
-        return Emulative::PHP_8_0;
+final class MatchTokenEmulator extends KeywordEmulator {
+    public function getPhpVersion(): PhpVersion {
+        return PhpVersion::fromComponents(8, 0);
     }
 
-    public function getKeywordString(): string
-    {
+    public function getKeywordString(): string {
         return 'match';
     }
 
-    public function getKeywordToken(): int
-    {
+    public function getKeywordToken(): int {
         return \T_MATCH;
     }
 }