+++ src/ncc/ThirdParty/nikic/PhpParser/Node/Expr.php
@@ -4,6 +4,5 @@ namespace ncc\ThirdParty\nikic\PhpParser\Node;
 
 use ncc\ThirdParty\nikic\PhpParser\NodeAbstract;
 
-abstract class Expr extends NodeAbstract
-{
+abstract class Expr extends NodeAbstract {
 }