+++ src/ncc/ThirdParty/nikic/PhpParser/Node/Expr/Array_.php
@@ -2,33 +2,33 @@
 
 namespace ncc\ThirdParty\nikic\PhpParser\Node\Expr;
 
+use ncc\ThirdParty\nikic\PhpParser\Node\ArrayItem;
 use ncc\ThirdParty\nikic\PhpParser\Node\Expr;
 
-class Array_ extends Expr
-{
+class Array_ extends Expr {
     // For use in "kind" attribute
-    const KIND_LONG = 1;  // array() syntax
-    const KIND_SHORT = 2; // [] syntax
+    public const KIND_LONG = 1;  // array() syntax
+    public const KIND_SHORT = 2; // [] syntax
 
-    /** @var (ArrayItem|null)[] Items */
-    public $items;
+    /** @var ArrayItem[] Items */
+    public array $items;
 
     /**
      * Constructs an array node.
      *
-     * @param (ArrayItem|null)[] $items      Items of the array
-     * @param array       $attributes Additional attributes
+     * @param ArrayItem[] $items Items of the array
+     * @param array<string, mixed> $attributes Additional attributes
      */
     public function __construct(array $items = [], array $attributes = []) {
         $this->attributes = $attributes;
         $this->items = $items;
     }
 
-    public function getSubNodeNames() : array {
+    public function getSubNodeNames(): array {
         return ['items'];
     }
-    
-    public function getType() : string {
+
+    public function getType(): string {
         return 'Expr_Array';
     }
 }