+++ src/ncc/ThirdParty/nikic/PhpParser/Node/Expr/AssignOp/BitwiseAnd.php
@@ -4,9 +4,8 @@ namespace ncc\ThirdParty\nikic\PhpParser\Node\Expr\AssignOp;
 
 use ncc\ThirdParty\nikic\PhpParser\Node\Expr\AssignOp;
 
-class BitwiseAnd extends AssignOp
-{
-    public function getType() : string {
+class BitwiseAnd extends AssignOp {
+    public function getType(): string {
         return 'Expr_AssignOp_BitwiseAnd';
     }
 }