+++ src/ncc/ThirdParty/nikic/PhpParser/Node/Expr/AssignOp/BitwiseOr.php
@@ -4,9 +4,8 @@ namespace ncc\ThirdParty\nikic\PhpParser\Node\Expr\AssignOp;
 
 use ncc\ThirdParty\nikic\PhpParser\Node\Expr\AssignOp;
 
-class BitwiseOr extends AssignOp
-{
-    public function getType() : string {
+class BitwiseOr extends AssignOp {
+    public function getType(): string {
         return 'Expr_AssignOp_BitwiseOr';
     }
 }