+++ src/ncc/ThirdParty/nikic/PhpParser/Node/Expr/AssignOp/Mul.php
@@ -4,9 +4,8 @@ namespace ncc\ThirdParty\nikic\PhpParser\Node\Expr\AssignOp;
 
 use ncc\ThirdParty\nikic\PhpParser\Node\Expr\AssignOp;
 
-class Mul extends AssignOp
-{
-    public function getType() : string {
+class Mul extends AssignOp {
+    public function getType(): string {
         return 'Expr_AssignOp_Mul';
     }
 }