+++ src/ncc/ThirdParty/nikic/PhpParser/Node/Expr/AssignOp/Plus.php
@@ -4,9 +4,8 @@ namespace ncc\ThirdParty\nikic\PhpParser\Node\Expr\AssignOp;
 
 use ncc\ThirdParty\nikic\PhpParser\Node\Expr\AssignOp;
 
-class Plus extends AssignOp
-{
-    public function getType() : string {
+class Plus extends AssignOp {
+    public function getType(): string {
         return 'Expr_AssignOp_Plus';
     }
 }