+++ src/ncc/ThirdParty/nikic/PhpParser/Node/Expr/BinaryOp/BitwiseAnd.php
@@ -4,13 +4,12 @@ namespace ncc\ThirdParty\nikic\PhpParser\Node\Expr\BinaryOp;
 
 use ncc\ThirdParty\nikic\PhpParser\Node\Expr\BinaryOp;
 
-class BitwiseAnd extends BinaryOp
-{
-    public function getOperatorSigil() : string {
+class BitwiseAnd extends BinaryOp {
+    public function getOperatorSigil(): string {
         return '&';
     }
-    
-    public function getType() : string {
+
+    public function getType(): string {
         return 'Expr_BinaryOp_BitwiseAnd';
     }
 }