+++ src/ncc/ThirdParty/nikic/PhpParser/Node/Expr/MethodCall.php
@@ -2,27 +2,27 @@
 
 namespace ncc\ThirdParty\nikic\PhpParser\Node\Expr;
 
+use ncc\ThirdParty\nikic\PhpParser\Node;
 use ncc\ThirdParty\nikic\PhpParser\Node\Arg;
 use ncc\ThirdParty\nikic\PhpParser\Node\Expr;
 use ncc\ThirdParty\nikic\PhpParser\Node\Identifier;
 use ncc\ThirdParty\nikic\PhpParser\Node\VariadicPlaceholder;
 
-class MethodCall extends CallLike
-{
+class MethodCall extends CallLike {
     /** @var Expr Variable holding object */
-    public $var;
+    public Expr $var;
     /** @var Identifier|Expr Method name */
-    public $name;
+    public Node $name;
     /** @var array<Arg|VariadicPlaceholder> Arguments */
-    public $args;
+    public array $args;
 
     /**
      * Constructs a function call node.
      *
-     * @param Expr                           $var        Variable holding object
-     * @param string|Identifier|Expr         $name       Method name
-     * @param array<Arg|VariadicPlaceholder> $args       Arguments
-     * @param array                          $attributes Additional attributes
+     * @param Expr $var Variable holding object
+     * @param string|Identifier|Expr $name Method name
+     * @param array<Arg|VariadicPlaceholder> $args Arguments
+     * @param array<string, mixed> $attributes Additional attributes
      */
     public function __construct(Expr $var, $name, array $args = [], array $attributes = []) {
         $this->attributes = $attributes;
@@ -31,11 +31,11 @@ class MethodCall extends CallLike
         $this->args = $args;
     }
 
-    public function getSubNodeNames() : array {
+    public function getSubNodeNames(): array {
         return ['var', 'name', 'args'];
     }
-    
-    public function getType() : string {
+
+    public function getType(): string {
         return 'Expr_MethodCall';
     }