+++ src/ncc/ThirdParty/nikic/PhpParser/Node/Expr/PostDec.php
@@ -4,27 +4,26 @@ namespace ncc\ThirdParty\nikic\PhpParser\Node\Expr;
 
 use ncc\ThirdParty\nikic\PhpParser\Node\Expr;
 
-class PostDec extends Expr
-{
+class PostDec extends Expr {
     /** @var Expr Variable */
-    public $var;
+    public Expr $var;
 
     /**
      * Constructs a post decrement node.
      *
-     * @param Expr  $var        Variable
-     * @param array $attributes Additional attributes
+     * @param Expr $var Variable
+     * @param array<string, mixed> $attributes Additional attributes
      */
     public function __construct(Expr $var, array $attributes = []) {
         $this->attributes = $attributes;
         $this->var = $var;
     }
 
-    public function getSubNodeNames() : array {
+    public function getSubNodeNames(): array {
         return ['var'];
     }
-    
-    public function getType() : string {
+
+    public function getType(): string {
         return 'Expr_PostDec';
     }
 }