+++ src/ncc/ThirdParty/nikic/PhpParser/Node/Expr/Throw_.php
@@ -4,27 +4,26 @@ namespace ncc\ThirdParty\nikic\PhpParser\Node\Expr;
 
 use ncc\ThirdParty\nikic\PhpParser\Node;
 
-class Throw_ extends Node\Expr
-{
+class Throw_ extends Node\Expr {
     /** @var Node\Expr Expression */
-    public $expr;
+    public Node\Expr $expr;
 
     /**
      * Constructs a throw expression node.
      *
-     * @param Node\Expr $expr       Expression
-     * @param array     $attributes Additional attributes
+     * @param Node\Expr $expr Expression
+     * @param array<string, mixed> $attributes Additional attributes
      */
     public function __construct(Node\Expr $expr, array $attributes = []) {
         $this->attributes = $attributes;
         $this->expr = $expr;
     }
 
-    public function getSubNodeNames() : array {
+    public function getSubNodeNames(): array {
         return ['expr'];
     }
 
-    public function getType() : string {
+    public function getType(): string {
         return 'Expr_Throw';
     }
 }