+++ src/ncc/ThirdParty/nikic/PhpParser/Node/Scalar/EncapsedStringPart.php
@@ -1,30 +1,3 @@
 <?php declare(strict_types=1);
 
-namespace ncc\ThirdParty\nikic\PhpParser\Node\Scalar;
-
-use ncc\ThirdParty\nikic\PhpParser\Node\Scalar;
-
-class EncapsedStringPart extends Scalar
-{
-    /** @var string String value */
-    public $value;
-
-    /**
-     * Constructs a node representing a string part of an encapsed string.
-     *
-     * @param string $value      String value
-     * @param array  $attributes Additional attributes
-     */
-    public function __construct(string $value, array $attributes = []) {
-        $this->attributes = $attributes;
-        $this->value = $value;
-    }
-
-    public function getSubNodeNames() : array {
-        return ['value'];
-    }
-    
-    public function getType() : string {
-        return 'Scalar_EncapsedStringPart';
-    }
-}
+require __DIR__ . '/../InterpolatedStringPart.php';