+++ src/ncc/ThirdParty/nikic/PhpParser/Node/Scalar/MagicConst/Dir.php
@@ -4,13 +4,12 @@ namespace ncc\ThirdParty\nikic\PhpParser\Node\Scalar\MagicConst;
 
 use ncc\ThirdParty\nikic\PhpParser\Node\Scalar\MagicConst;
 
-class Dir extends MagicConst
-{
-    public function getName() : string {
+class Dir extends MagicConst {
+    public function getName(): string {
         return '__DIR__';
     }
-    
-    public function getType() : string {
+
+    public function getType(): string {
         return 'Scalar_MagicConst_Dir';
     }
 }