+++ src/ncc/ThirdParty/nikic/PhpParser/Node/Scalar/MagicConst/Function_.php
@@ -4,13 +4,12 @@ namespace ncc\ThirdParty\nikic\PhpParser\Node\Scalar\MagicConst;
 
 use ncc\ThirdParty\nikic\PhpParser\Node\Scalar\MagicConst;
 
-class Function_ extends MagicConst
-{
-    public function getName() : string {
+class Function_ extends MagicConst {
+    public function getName(): string {
         return '__FUNCTION__';
     }
-    
-    public function getType() : string {
+
+    public function getType(): string {
         return 'Scalar_MagicConst_Function';
     }
 }