+++ src/ncc/ThirdParty/nikic/PhpParser/Node/Stmt.php
@@ -4,6 +4,5 @@ namespace ncc\ThirdParty\nikic\PhpParser\Node;
 
 use ncc\ThirdParty\nikic\PhpParser\NodeAbstract;
 
-abstract class Stmt extends NodeAbstract
-{
+abstract class Stmt extends NodeAbstract {
 }