+++ src/ncc/ThirdParty/nikic/PhpParser/Node/Stmt/InlineHTML.php
@@ -4,27 +4,26 @@ namespace ncc\ThirdParty\nikic\PhpParser\Node\Stmt;
 
 use ncc\ThirdParty\nikic\PhpParser\Node\Stmt;
 
-class InlineHTML extends Stmt
-{
+class InlineHTML extends Stmt {
     /** @var string String */
-    public $value;
+    public string $value;
 
     /**
      * Constructs an inline HTML node.
      *
-     * @param string $value      String
-     * @param array  $attributes Additional attributes
+     * @param string $value String
+     * @param array<string, mixed> $attributes Additional attributes
      */
     public function __construct(string $value, array $attributes = []) {
         $this->attributes = $attributes;
         $this->value = $value;
     }
 
-    public function getSubNodeNames() : array {
+    public function getSubNodeNames(): array {
         return ['value'];
     }
-    
-    public function getType() : string {
+
+    public function getType(): string {
         return 'Stmt_InlineHTML';
     }
 }