+++ src/ncc/ThirdParty/nikic/PhpParser/Parser.php
@@ -2,8 +2,7 @@
 
 namespace ncc\ThirdParty\nikic\PhpParser;
 
-interface Parser
-{
+interface Parser {
     /**
      * Parses PHP code into a node tree.
      *
@@ -11,8 +10,15 @@ interface Parser
      * @param ErrorHandler|null $errorHandler Error handler to use for lexer/parser errors, defaults
      *                                        to ErrorHandler\Throwing.
      *
-     * @return Node\Stmt[]|null Array of statements (or null non-throwing error handler is used and
+     * @return ncc\ThirdParty\nikic\PhpParser\Node\Stmt[]|null Array of statements (or null non-throwing error handler is used and
      *                          the parser was unable to recover from an error).
      */
-    public function parse(string $code, ErrorHandler $errorHandler = null);
+    public function parse(string $code, ?ErrorHandler $errorHandler = null): ?array;
+
+    /**
+     * Return tokens for the last parse.
+     *
+     * @return Token[]
+     */
+    public function getTokens(): array;
 }